requirements.txt
numpy>=1.19.5               # 1.22.4
pandas==1.3.5
scipy>=1.6.1                # 1.10.1
scikit-learn==1.0.2
xgboost==1.5.2
lightgbm>=3.0.0             # 3.3.5
catboost>=0.24.4            # 1.2
joblib>=1.0.1               # 1.2.0
tabulate>=0.8.7             # 0.9.0
matplotlib>=3.2.2           # 3.7.1
dtreeviz==2.0.0
numba==0.56.4
shap>=0.36.0                # 0.41.0
seaborn>=0.11.1             # 0.12.2
wordcloud>=1.8.1            # 1.9.2
category_encoders>=2.2.2    # 2.6.1
optuna>=2.7.0               # 3.1.1
scikit-plot==0.3.7
markdown                    # 3.4.3
typing-extensions           # 4.6.1
mljar-supervised==0.11.5
streamlit==1.22.0
altair==4.2.2
ipython                     # 8.13.2

# testing requirements
pytest==7.3.1
locust==2.15.1
selenium==4.9.1
seleniumbase==4.14.12
opencv-python==4.7.0.72

# dev requirements
pdoc                        # 13.1.1
pre-commit                  # 3.3.2
